requirements.txt
slack-bolt==1.16.2
slack-sdk==3.20.0
python-dotenv~=1.0.0
